server diversification
